+++ src/config/bot.ts
@@ -1,5 +1,5 @@
 export const bot = {
   name: 'Tatiana',
-  version: '2.1.0',
+  version: '2.2.0',
   creatorName: 'Enkidu',
 } as const;